backend/requirements.txt
flask
flask-cors
requests
numpy<2
python-docx
flask-jwt-extended
python-dotenv
bcrypt